moved log out button into the right sidebar/menu
