+++ src/components/header/header.js
@@ -1,18 +1,18 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { Header, Image, Menu, Button, Icon } from 'semantic-ui-react';
-import { connect } from 'react-redux';
-import { push } from 'react-router-redux';
-import { logout } from 'redux/modules/user';
+// import { connect } from 'react-redux';
+// import { push } from 'react-router-redux';
+// import { logout } from 'redux/modules/user';
 import Logo from 'assets/svg/is_tab_black.svg';
 import './style/header.scss';
 
-class DashHeader extends Component {
+export class DashHeader extends Component {
   static propTypes = {
     title: PropTypes.string.isRequired,
     callback: PropTypes.func,
     history: PropTypes.func, // eslint-disable-line
-    logout: PropTypes.func,
+    // logout: PropTypes.func,
     toggleRightSidebar: PropTypes.func
   };
 
@@ -65,12 +65,6 @@ class DashHeader extends Component {
             {this.props.title}
           </Header>
         </Menu.Item>
-        <Menu.Item style={{ marginLeft: '1rem' }} onClick={this.props.logout}>
-          <Header as='h3'>
-            <Icon name='log out' size='tiny'/>
-            Log out
-          </Header>
-        </Menu.Item>
         <Menu.Item onClick={this.props.toggleRightSidebar}>
           <Header as='h3'>
             <Icon name='tasks' size='tiny'/>
@@ -81,22 +75,3 @@ class DashHeader extends Component {
     );
   }
 }
-
-// const mapStateToProps = (state) => ({
-//   modal: state.modal.modal
-// });
-
-const mapDispatchToProps = dispatch => ({
-  logout: () => {
-    dispatch(logout());
-    dispatch(push('login'));
-  }
-});
-
-const DashHeaderContainer =
-  connect(
-    null,
-    mapDispatchToProps,
-  )(DashHeader);
-
-export { DashHeaderContainer as DashHeader };
